test_temp.py -> test_ModelParameters.py; PEP8
